Licence is BSD and minor changes
The site indicates BSD licence, (I created an issue to see if they (he) can add an explicit LICENCE file)
I removed a few things
The change log indicates an old commits, so I added a line with the new commit and incremented the version.  I don't know if it is right.